+++ tests/framework/db/ConnectionTest.php
@@ -88,6 +88,9 @@ class ConnectionTest extends DatabaseTestCase
         $this->assertEquals('`table`.`column`', $connection->quoteColumnName('table.`column`'));
         $this->assertEquals('[[column]]', $connection->quoteColumnName('[[column]]'));
         $this->assertEquals('{{column}}', $connection->quoteColumnName('{{column}}'));
+        $this->assertEquals('{{table}}.`column`', $connection->quoteColumnName('{{table}}.column'));
+        $this->assertEquals('{{table}}.`column`', $connection->quoteColumnName('{{table}}.`column`'));
+        $this->assertEquals('{{table}}.[[column]]', $connection->quoteColumnName('{{table}}.[[column]]'));
         $this->assertEquals('(column)', $connection->quoteColumnName('(column)'));
     }