+++ tests/framework/db/mssql/MssqlConnectionTest.php
@@ -40,6 +40,9 @@ class MssqlConnectionTest extends ConnectionTest
         $this->assertEquals('[table].[column]', $connection->quoteColumnName('table.[column]'));
         $this->assertEquals('[[column]]', $connection->quoteColumnName('[[column]]'));
         $this->assertEquals('{{column}}', $connection->quoteColumnName('{{column}}'));
+        $this->assertEquals('{{table}}.[column]', $connection->quoteColumnName('{{table}}.column'));
+        $this->assertEquals('{{table}}.[column]', $connection->quoteColumnName('{{table}}.[column]'));
+        $this->assertEquals('{{table}}.[[column]]', $connection->quoteColumnName('{{table}}.[[column]]'));
         $this->assertEquals('(column)', $connection->quoteColumnName('(column)'));
     }
 }